# artifacts/api-server/src/routes/payments.ts
import { Router } from "express";
import { eq, desc } from "drizzle-orm";
import { db, paymentsTable, subscriptionsTable, adminSettingsTable, bindingsTable, slaveAccountsTable } from "@workspace/db";
import { InitiatePaymentBody } from "@workspace/api-zod";
import { authenticate } from "../middlewares/authenticate";
import { logger } from "../lib/logger";

const router = Router();

async function getDailyFee(): Promise<number> {
  const [settings] = await db.select().from(adminSettingsTable).limit(1);
  return settings ? parseFloat(settings.dailyFee as string) : 100;
}

function addTradingDays(start: Date, days: number): Date {
  const result = new Date(start);
  let added = 0;
  while (added < days) {
    result.setDate(result.getDate() + 1);
    const day = result.getDay();
    if (day !== 0 && day !== 6) added++;
  }
  return result;
}

router.get("/payments", authenticate, async (req, res): Promise<void> => {
  const payments = await db
    .select()
    .from(paymentsTable)
    .where(eq(paymentsTable.userId, req.userId!))
    .orderBy(desc(paymentsTable.createdAt));

  res.json(
    payments.map((p) => ({
      ...p,
      amount: parseFloat(p.amount as string),
    }))
  );
});

router.post("/payments", authenticate, async (req, res): Promise<void> => {
  const parsed = InitiatePaymentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { phone, days } = parsed.data;
  const dailyFee = await getDailyFee();
  const amount = days * dailyFee;

  // Format phone: strip leading 0 or +254, normalize to 254XXXXXXXXX
  const normalizedPhone = phone.replace(/^\+/, "").replace(/^0/, "254");

  // Check if MPESA credentials are configured
  const consumerKey = process.env.MPESA_CONSUMER_KEY;
  const consumerSecret = process.env.MPESA_CONSUMER_SECRET;
  const passkey = process.env.MPESA_PASSKEY;
  const shortcode = process.env.MPESA_SHORTCODE;
  const callbackUrl = process.env.MPESA_CALLBACK_URL;

  if (!consumerKey || !consumerSecret || !passkey || !shortcode) {
    // Demo mode: simulate payment completion
    logger.warn("MPESA credentials not configured, using demo mode");

    const [payment] = await db
      .insert(paymentsTable)
      .values({
        userId: req.userId!,
        phone: normalizedPhone,
        amount: amount.toFixed(2),
        status: "completed",
        days,
        mpesaReceipt: `DEMO${Date.now()}`,
        checkoutRequestId: `DEMO-${Date.now()}`,
      })
      .returning();

    // Activate subscription
    await activateSubscription(req.userId!, days);

    res.json({
      checkoutRequestId: payment.checkoutRequestId!,
      message: "Demo mode: Payment simulated and subscription activated",
      amount,
    });
    return;
  }

  try {
    // Get M-Pesa access token
    const authResponse = await fetch(
      "https://sandbox.safaricom.co.ke/oauth/v1/generate?grant_type=client_credentials",
      {
        headers: {
          Authorization: `Basic ${Buffer.from(`${consumerKey}:${consumerSecret}`).toString("base64")}`,
        },
      }
    );
    const authData = (await authResponse.json()) as { access_token: string };
    const accessToken = authData.access_token;

    const timestamp = new Date()
      .toISOString()
      .replace(/[-:T.Z]/g, "")
      .slice(0, 14);
    const password = Buffer.from(`${shortcode}${passkey}${timestamp}`).toString("base64");

    const stkResponse = await fetch("https://sandbox.safaricom.co.ke/mpesa/stkpush/v1/processrequest", {
      method: "POST",
      headers: {
        Authorization: `Bearer ${accessToken}`,
        "Content-Type": "application/json",
      },
      body: JSON.stringify({
        BusinessShortCode: shortcode,
        Password: password,
        Timestamp: timestamp,
        TransactionType: "CustomerPayBillOnline",
        Amount: Math.ceil(amount),
        PartyA: normalizedPhone,
        PartyB: shortcode,
        PhoneNumber: normalizedPhone,
        CallBackURL: callbackUrl,
        AccountReference: "PESAMATRIX",
        TransactionDesc: `${days} trading day(s) subscription`,
      }),
    });

    const stkData = (await stkResponse.json()) as { CheckoutRequestID?: string; errorMessage?: string };

    if (!stkData.CheckoutRequestID) {
      res.status(400).json({ error: stkData.errorMessage ?? "STK Push failed" });
      return;
    }

    const [payment] = await db
      .insert(paymentsTable)
      .values({
        userId: req.userId!,
        phone: normalizedPhone,
        amount: amount.toFixed(2),
        status: "pending",
        days,
        checkoutRequestId: stkData.CheckoutRequestID,
      })
      .returning();

    res.json({
      checkoutRequestId: payment.checkoutRequestId!,
      message: "STK Push sent to your phone. Complete the payment to activate subscription.",
      amount,
    });
  } catch (err) {
    req.log.error({ err }, "STK Push error");
    res.status(500).json({ error: "Payment initiation failed" });
  }
});

router.post("/payments/callback", async (req, res): Promise<void> => {
  try {
    const body = req.body?.Body?.stkCallback;
    if (!body) {
      res.json({ message: "OK" });
      return;
    }

    const { CheckoutRequestID, ResultCode, CallbackMetadata } = body;

    const [payment] = await db
      .select()
      .from(paymentsTable)
      .where(eq(paymentsTable.checkoutRequestId, CheckoutRequestID));

    if (!payment) {
      res.json({ message: "OK" });
      return;
    }

    if (ResultCode === 0) {
      // Success
      const receiptItem = CallbackMetadata?.Item?.find((i: { Name: string }) => i.Name === "MpesaReceiptNumber");
      const mpesaReceipt = receiptItem?.Value as string | undefined;

      await db
        .update(paymentsTable)
        .set({ status: "completed", mpesaReceipt })
        .where(eq(paymentsTable.id, payment.id));

      await activateSubscription(payment.userId, payment.days);
    } else {
      await db.update(paymentsTable).set({ status: "failed" }).where(eq(paymentsTable.id, payment.id));
    }

    res.json({ message: "OK" });
  } catch (err) {
    logger.error({ err }, "M-Pesa callback error");
    res.json({ message: "OK" });
  }
});

async function activateSubscription(userId: number, days: number): Promise<void> {
  const [existing] = await db.select().from(subscriptionsTable).where(eq(subscriptionsTable.userId, userId));

  const now = new Date();

  function addTradingDaysLocal(start: Date, d: number): Date {
    return addTradingDays(start, d);
  }

  if (!existing) {
    const endDate = addTradingDaysLocal(now, days);
    await db.insert(subscriptionsTable).values({
      userId,
      status: "active",
      startDate: now,
      endDate,
      daysPaid: days,
    });
  } else {
    // Extend from current end date if still active, otherwise from now
    const baseDate = existing.status === "active" && existing.endDate && existing.endDate > now
      ? existing.endDate
      : now;
    const newEndDate = addTradingDaysLocal(baseDate, days);

    await db
      .update(subscriptionsTable)
      .set({
        status: "active",
        startDate: existing.startDate ?? now,
        endDate: newEndDate,
        daysPaid: existing.daysPaid + days,
      })
      .where(eq(subscriptionsTable.userId, userId));

    // Reactivate suspended bindings
    const userSlaveAccounts = await db
      .select()
      .from(slaveAccountsTable)
      .where(eq(slaveAccountsTable.userId, userId));

    for (const slave of userSlaveAccounts) {
      await db
        .update(bindingsTable)
        .set({ status: "active" })
        .where(eq(bindingsTable.slaveAccountId, slave.id));
    }
  }
}

export default router;