# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # SUBSCRIPTIONS
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List current user's payment history
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate M-Pesa STK Push payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: STK Push initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/callback:
    post:
      operationId: mpesaCallback
      tags: [payments]
      summary: M-Pesa callback endpoint (called by Safaricom)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpesaCallbackInput"
      responses:
        "200":
          description: Callback processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # MASTER ACCOUNTS
...
  /slave-accounts:
    get:
      operationId: listSlaveAccounts
      tags: [slaveAccounts]
      summary: List current user's slave accounts
      responses:
        "200":
          description: List of slave accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SlaveAccount"
    post:
      operationId: createSlaveAccount
      tags: [slaveAccounts]
      summary: Create a new slave MT5 account via MetaApi CopyFactory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlaveAccountInput"
      responses:
        "201":
          description: Slave account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaveAccount"
        "400":
          description: Invalid request or subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /slave-accounts/{id}:
    delete:
      operationId: deleteSlaveAccount
      tags: [slaveAccounts]
      summary: Delete a slave account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # STRATEGIES
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List current user's strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new CopyFactory strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /strategies/{id}:
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # BINDINGS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    User:
      type: object
      required: [id, name, email, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 6

    # SUBSCRIPTION
...
    PaymentInput:
      type: object
      required: [phone, days]
      properties:
        phone:
          type: string
        days:
          type: integer
          minimum: 1
          maximum: 365
    PaymentInitResponse:
      type: object
      required: [checkoutRequestId, message]
      properties:
        checkoutRequestId:
          type: string
        message:
          type: string
        amount:
          type: number
    MpesaCallbackInput:
      type: object
      properties:
        Body:
          type: object
    Payment:
      type: object
      required: [id, userId, phone, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        phone:
          type: string
        amount:
          type: number
        mpesaReceipt:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed]
        days:
          type: integer
        createdAt:
          type: string
          format: date-time

    # MASTER ACCOUNT
...
    SlaveAccountInput:
      type: object
      required: [broker, server, mt5Login, investorPassword]
      properties:
        broker:
          type: string
        server:
          type: string
        mt5Login:
          type: string
        investorPassword:
          type: string
    SlaveAccount:
      type: object
      required: [id, userId, mt5Login, broker, server, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        metaapiAccountId:
          type: ["string", "null"]
        subscriberId:
          type: ["string", "null"]
        mt5Login:
          type: string
        broker:
          type: string
        server:
          type: string
        status:
          type: string
          enum: [connecting, connected, disconnected, suspended, error]
        createdAt:
          type: string
          format: date-time

    # STRATEGY
    StrategyInput:
      type: object
      required: [strategyName, masterAccountId]
      properties:
        strategyName:
          type: string
          minLength: 2
        masterAccountId:
          type: integer
    Strategy:
      type: object
      required: [id, userId, strategyName, masterAccountId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        copyfactoryStrategyId:
          type: ["string", "null"]
        strategyName:
          type: string
        masterAccountId:
          type: integer
        status:
          type: string
          enum: [active, paused, error]
        createdAt:
          type: string
          format: date-time

    # BINDING
...