# lib/api-zod/src/generated/types/slaveAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SlaveAccountStatus } from './slaveAccountStatus';

export interface SlaveAccount {
  id: number;
  userId: number;
  /** @nullable */
  metaapiAccountId?: string | null;
  /** @nullable */
  subscriberId?: string | null;
  mt5Login: string;
  broker: string;
  server: string;
  status: SlaveAccountStatus;
  createdAt: Date;
}